requirements.txt
Django==2.2.0
djangorestframework>=3.9.0,<3.10.0
django-extensions>=3.1.1
psycopg2>=2.7.5,<2.8.0
marshmallow
